+++ simple-stack/src/test/java/com/zhuinden/simplestack/TestSuite.java
@@ -7,6 +7,6 @@ import org.junit.runners.Suite;
  * Created by Owner on 2017. 01. 17..
  */
 @RunWith(Suite.class)
-@Suite.SuiteClasses({ExampleUnitTest.class, FlowTest.class, ReentranceTest.class, BackstackTest.class, HistoryBuilderTest.class})
+@Suite.SuiteClasses({ExampleUnitTest.class, FlowTest.class, ReentranceTest.class, BackstackTest.class, HistoryBuilderTest.class, BackstackDelegateTest.class})
 public class TestSuite {
 }